piatrika_forms_restapi/piatrika_forms_restapi/requirements-dev.txt
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
mysqlclient==1.4.2.post1
Pillow==6.0.0
PyMySQL==0.9.3
pytz==2019.1
sqlparse==0.3.0
yapf==0.27.0
